requirements.txt
altgraph==0.17
asgiref==3.3.4
certifi==2020.12.5
click==7.1.2
Django==3.2.3
future==0.18.2
image==1.5.33
importlib-metadata==4.3.1
mkl-fft==1.3.0
mkl-random
mkl-service==2.3.0
numpy
olefile==0.46
opencv-python==4.5.2.52
pefile==2021.5.24
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pytesseract==0.3.7
python-dotenv==0.17.1
pytz==2021.1
pywin32==227
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
wincertstore==0.2
zipp==3.4.1
